Changed WordRange of the initial empty hypothesis to be [-1..-1].
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@1733 1f5c12ca-751b-0410-a591-d2e778427230

+++ branches/mtm2_cube_pruning/moses/src/WordsRange.h
@@ -32,19 +32,19 @@ class WordsRange
 {
 	friend std::ostream& operator << (std::ostream& out, const WordsRange& range);
 
-	size_t m_startPos, m_endPos;
+	int m_startPos, m_endPos;
 public:
-	inline WordsRange(size_t startPos, size_t endPos) : m_startPos(startPos), m_endPos(endPos) {}
+	inline WordsRange(int startPos, int endPos) : m_startPos(startPos), m_endPos(endPos) {}
 	inline WordsRange(const WordsRange &copy)
 	 : m_startPos(copy.GetStartPos())
 	 , m_endPos(copy.GetEndPos())
 	 {}
 	
-	inline size_t GetStartPos() const
+	inline int GetStartPos() const
 	{
 		return m_startPos;
 	}
-	inline size_t GetEndPos() const
+	inline int GetEndPos() const
 	{
 		return m_endPos;
 	}
